+++ pkg/core/evaluator/eval_calc_test.go
@@ -1,0 +1,87 @@
+package evaluator
+
+import (
+	"github.com/raa0121/GoBCDice/pkg/core/dice/feeder"
+	"github.com/raa0121/GoBCDice/pkg/core/dice/roller"
+	"github.com/raa0121/GoBCDice/pkg/core/object"
+	"github.com/raa0121/GoBCDice/pkg/core/parser"
+	"testing"
+)
+
+func TestEvalCalc(t *testing.T) {
+	testcases := []struct {
+		input    string
+		expected int
+	}{
+		{"C(5)", 5},
+		{"C(10)", 10},
+		{"C(42)", 42},
+		{"C(65535)", 65535},
+		{"C(-5)", -5},
+		{"C(-10)", -10},
+		{"C(-42)", -42},
+		{"C(-65535)", -65535},
+		{"C(+5)", 5},
+		{"C(+10)", 10},
+		{"C(+42)", 42},
+		{"C(+65535)", 65535},
+		{"C(1+2)", 3},
+		{"C(1-2)", -1},
+		{"C(1*2)", 2},
+		{"C(1/2)", 0},
+		{"C(-1+2)", 1},
+		{"C(+1+2)", 3},
+		{"C(1+2-3)", 0},
+		{"C(1*2+3)", 5},
+		{"C(1/2+3)", 3},
+		{"C(1+2*3)", 7},
+		{"C(1+2/3)", 1},
+		{"C(1+(2-3))", 0},
+		{"C((1+2)*3)", 9},
+		{"C((1+2)/3)", 1},
+		{"C(-(1+2))", -3},
+		{"C(+(1+2))", 3},
+		{"C(1+50/3-2)", 15},
+		{"C(1+50/3u-2)", 16},
+		{"C(1+50/3r-2)", 16},
+		{"C(1+100/3-2)", 32},
+		{"C(1+100/3u-2)", 33},
+		{"C(1+100/3r-2)", 32},
+	}
+
+	for _, test := range testcases {
+		t.Run(test.input, func(t *testing.T) {
+			ast, parseErr := parser.Parse(test.input)
+			if parseErr != nil {
+				t.Errorf("構文解析エラー: %s", parseErr)
+				return
+			}
+
+			// ノードを評価する
+			dieFeeder := feeder.NewEmptyQueue()
+			evaluator := NewEvaluator(roller.New(dieFeeder), NewEnvironment())
+
+			evaluated, evalErr := evaluator.Eval(ast)
+			if evalErr != nil {
+				t.Errorf("評価エラー: %s", evalErr)
+				return
+			}
+
+			if evaluated == nil {
+				t.Error("Evalの対象外 (nil)")
+				return
+			}
+
+			// 型が合っているか？
+			obj, typeMatched := evaluated.(*object.Integer)
+			if !typeMatched {
+				t.Errorf("整数オブジェクトでない: %T (%+v)", obj, obj)
+				return
+			}
+
+			if obj.Value != test.expected {
+				t.Errorf("異なる値: got=%d, want=%d", obj.Value, test.expected)
+			}
+		})
+	}
+}
